:rotating_light: PATHFINDING PATHFINDING :rotating_light:
The LocalTransferGrid now has the ability to iteratively cache discovered pathways between every connector that is attached to any external energy producer or consumer. Connectors that don't participate in passing power into or out of the network are ignored. The pathways are derived from the A* heuristic approach, using the Euclidian distance, and will soon prioritize pathways/wires with lower impedance (Ohms)

Functional power transfer is very close.

Some specifics:
- GridEdges now store a direct reference to their WireSpool
- GridEdges also store a heuristic comprised of the Euclidian distance between their start and end block positions
- GridPaths can be instantiated from Sets
- GridVertices can calculate an arbitrary heuristic from themselves to a given GridVertex
- GridVertices keep track of the F, G, and H values commonly associated with A*
- GridClientCache recursive StackOverflow check is unncessary and caused issues so it's been removed
- Trimmed unused overloads and helpers from GlobalTransferGrid
- Some final renames and comments to bring old classes up to speed
- Barebones framework is in place to (eventually idk) transition the basic ForgeEnergy protocol over to implicit Kilowatts